% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_air.r
\name{check_air}
\alias{check_air}
\title{Check \code{air} setup}
\usage{
check_air()
}
\value{
\code{NULL} invisibly.  As a side effect prints a message.
}
\description{
\code{check_air()} checks the \code{air} setup
}
\examples{
\dontrun{
  check_air()
}
}
\seealso{
\code{\link[=use_tld_air]{use_tld_air()}}
}
