% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tld_all.R
\name{check_all}
\alias{check_all}
\title{Run all checks contained in this package}
\usage{
check_all()
}
\value{
\code{NULL} invisibly.  As a side effect prints several messages.
}
\description{
\code{check_tld_all()} runs all the checks contained in this package.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  check_all()
}
}
\seealso{
\code{\link[=use_tld_all]{use_tld_all()}}
}
