% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filenames.R
\name{check_filenames}
\alias{check_filenames}
\title{Check R file extensions}
\usage{
check_filenames()
}
\value{
\code{NULL} invisibly.  As a side effect prints a message.
}
\description{
\code{check_filenames()} checks that all R files use \code{.R} extension instead of \code{.r}.
}
\examples{
\dontrun{
  check_filenames()
}
}
\seealso{
\code{\link[=use_tld_filenames]{use_tld_filenames()}}
}
