% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_github_actions.R
\name{check_github_actions}
\alias{check_github_actions}
\title{Check GitHub Actions setup}
\usage{
check_github_actions()
}
\value{
\code{NULL} invisibly.  As a side effect prints a message.
}
\description{
\code{check_github_actions()} checks the GitHub Actions setup
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  check_github_actions()
}
}
\seealso{
\code{\link[=use_tld_github_actions]{use_tld_github_actions()}}
}
