% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_github_labels.r
\name{check_github_labels}
\alias{check_github_labels}
\title{Check GitHub labels}
\usage{
check_github_labels()
}
\value{
\code{NULL} invisibly.  As a side effect prints a message.
}
\description{
\code{check_github_labels()} checks GitHub labels.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  check_github_labels()
}
}
\seealso{
\code{\link[=use_tld_github_labels]{use_tld_github_labels()}}
}
