% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_github_settings.r
\name{check_github_settings}
\alias{check_github_settings}
\title{Check GitHub settings}
\usage{
check_github_settings()
}
\value{
\code{NULL} invisibly.  As a side effect prints a message.
}
\description{
\code{check_github_settings()} checks GitHub settings
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  check_github_settings()
}
}
\seealso{
\code{\link[=use_tld_github_settings]{use_tld_github_settings()}}
}
