% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_air.r
\name{use_tld_air}
\alias{use_tld_air}
\title{Use Trevor's preferred \code{air} settings}
\usage{
use_tld_air()
}
\value{
\code{NULL} invisibly.  As a side effect writes several files.
}
\description{
\code{use_tld_air()} uses Trevor's preferred \code{air} settings.
}
\examples{
\dontrun{
  use_tld_air()
}
}
\seealso{
\code{\link[=check_air]{check_air()}}
}
