% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tld_all.r
\name{use_tld_all}
\alias{use_tld_all}
\title{Use all of Trevor's preferred setups}
\usage{
use_tld_all()
}
\value{
\code{NULL} invisibly.  As a side effect writes several files.
}
\description{
\code{use_tld_all()} uses all of Trevor's setup functions in this package.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  use_tld_all()
}
}
\seealso{
\code{\link[=check_all]{check_all()}}
}
