% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filenames.R
\name{use_tld_filenames}
\alias{use_tld_filenames}
\title{Rename \code{.r} files to \code{.R}}
\usage{
use_tld_filenames()
}
\value{
\code{NULL} invisibly.  As a side effect renames files and runs \code{devtools::document()}.
}
\description{
\code{use_tld_filenames()} renames all R files ending with \code{.r} to use \code{.R} instead.
}
\examples{
\dontrun{
  use_tld_filenames()
}
}
\seealso{
\code{\link[=check_filenames]{check_filenames()}}
}
