% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_github_actions.r
\name{use_tld_github_actions}
\alias{use_tld_github_actions}
\title{Use Trevor's preferred GitHub Actions}
\usage{
use_tld_github_actions()
}
\value{
\code{NULL} invisibly.  As a side effect writes several files.
}
\description{
\code{use_tld_github_actions()} uses Trevor's preferred GitHub Actions
}
\examples{
\dontrun{
  use_tld_github_actions()
}
}
\seealso{
\code{\link[=check_github_actions]{check_github_actions()}}
}
