% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_github_labels.R
\name{use_tld_github_labels}
\alias{use_tld_github_labels}
\title{Use Trevor's preferred GitHub labels}
\usage{
use_tld_github_labels()
}
\value{
\code{NULL} invisibly.  As a side effect edits GitHub labels
}
\description{
\code{use_tld_github_labels()} uses Trevor's preferred GitHub labels.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  use_tld_github_labels()
}
}
\seealso{
\code{\link[=check_github_labels]{check_github_labels()}}
}
