% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_github_settings.r
\name{use_tld_github_settings}
\alias{use_tld_github_settings}
\title{Use Trevor's preferred GitHub settings}
\usage{
use_tld_github_settings(
  ...,
  delete_branch_on_merge = TRUE,
  enable_discussions = FALSE,
  enable_issues = TRUE,
  enable_merge_commit = FALSE,
  enable_projects = FALSE,
  enable_rebase_merge = TRUE,
  enable_squash_merge = TRUE,
  enable_wiki = FALSE
)
}
\arguments{
\item{...}{Passed to \code{\link[ghcli:gh_repo_edit]{ghcli::gh_repo_edit()}}}

\item{delete_branch_on_merge}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_discussions}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_issues}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_merge_commit}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_projects}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_rebase_merge}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_squash_merge}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_wiki}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}
}
\value{
\code{NULL} invisibly.  As a side effect edits GitHub settings.
}
\description{
\code{use_tld_github_settings()} uses Trevor's preferred GitHub settings.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  use_tld_github_settings()
}
}
\seealso{
\code{\link[=check_github_settings]{check_github_settings()}}
}
