% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_github_sponsors.R
\name{use_tld_github_sponsors}
\alias{use_tld_github_sponsors}
\title{Use Trevor's preferred GitHub sponsor button settings}
\usage{
use_tld_github_sponsors()
}
\value{
\code{NULL} invisibly.  As a side effect writes a file.
}
\description{
\code{use_tld_github_sponsors()} uses Trevor's preferred GitHub sponsor button settings.
}
\examples{
\dontrun{
  use_tld_github_sponsors()
}
}
\seealso{
\code{\link[=check_github_sponsors]{check_github_sponsors()}}
}
