% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_air.r
\name{use_tld_Rbuildignore}
\alias{use_tld_Rbuildignore}
\title{Use Trevor's preferred \code{.Rbuildignore} file}
\usage{
use_tld_Rbuildignore()
}
\value{
\code{NULL} invisibly.  As a side effect write a file.
}
\description{
\code{use_tld_Rbuildignore()} uses Trevor's preferred \code{.Rbuildignore} file.
}
\examples{
\dontrun{
  use_tld_Rbuildignore()
}
}
