% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_github_sponsors.R
\name{check_github_sponsors}
\alias{check_github_sponsors}
\title{Check GitHub sponsor button setup}
\usage{
check_github_sponsors()
}
\value{
\code{NULL} invisibly.  As a side effect prints a message.
}
\description{
\code{check_github_sponsors()} checks the GitHub sponsor button setup.
}
\examples{
\dontrun{
  check_github_sponsors()
}
}
\seealso{
\code{\link[=use_tld_github_sponsors]{use_tld_github_sponsors()}}
}
